refactored codebase, prepping for modularization of AstroPotography.py to refactor code architecture
