+++ pyproject.toml
@@ -1,0 +1,84 @@
+[build-system]
+requires      = ["setuptools>=61.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "astropitography"
+version = "1.0.10"
+description = "Astrophotography + raspberry pi HQ camera = AstroPitography"
+readme = "README.md"
+requires-python = ">=3.7"
+keywords = ["astrophotography", "raspberry pi"]
+authors = [
+    { name = "Adam Baskerville", email = "adamlukebaskerville@gmail.com" },
+]
+maintainers = [
+  { name = "Adam Baskerville", email = "adamlukebaskerville@gmail.com" },
+  { name = "Daniel Opara", email = "daniel.opara@tufts.edu" }
+]
+license = { file = "LICENSE" }
+classifiers = [
+  "Development Status :: 5 - Production/Stable",
+  "Intended Audience :: Developers",
+  "Intended Audience :: Science/Research",
+  "Intended Audience :: Education",
+
+  "Topic :: Education",
+  "Topic :: Multimedia :: Graphics :: Capture :: Digital Camera",
+  "Topic :: Multimedia :: Video :: Capture",
+  "Topic :: Scientific/Engineering :: Astronomy",
+  "Topic :: Scientific/Engineering :: Image Processing",
+  "Topic :: Software Development :: Build Tools",
+
+  "License :: OSI Approved :: MIT License",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3 :: Only",
+]
+
+dependencies = [
+  "scipy==1.8.1",
+  "numpy==1.24.3",
+  "picamera>=1.12",
+  "pidng>=3.4.7",
+  "Pillow>=8.4.0",
+  "PySimpleGUI>=4.55.1",
+]
+
+[project.optional-dependencies]
+dev = ["pip-tools", "pytest", "bumpver"]
+
+[project.urls]
+Homepage = "https://adambaskerville.github.io/tabs/astro/"
+
+[tool.setuptools]
+include-package-data = true
+
+[tool.setuptools.packages.find]
+namespaces = true
+where = ["src"]
+exclude = ["stack_test*", "tetra3*", "output*", "Old*", "images*", "AstroPitography_GUI.png", "__pycache__*"]
+
+[tool.setuptools.package-data]
+"astropitography.db" = ["*.npz"]
+"astropitography" = ["*.png"]
+
+[project.scripts]
+astropitography = "astropitography.__main__:main"
+
+[tool.bumpver]
+current_version = "1.0.5"
+version_pattern = "MAJOR.MINOR.PATCH[PYTAGNUM]"
+commit_message = "bump version {old_version} -> {new_version}"
+commit = true
+tag = true
+push = false
+
+[tool.bumpver.file_patterns]
+"pyproject.toml" = [
+    'current_version = "{version}"',
+]
+"README.md" = [
+    "{version}",
+    "{pep440_version}",
+]
+
