+++ src/astropitography/__main__.py
@@ -1,0 +1,7 @@
+from astropitography.AstroPitography import run
+
+def main():
+    run()
+
+if __name__ == "__main__":
+    main()
