--- tetra3/.gitignore
@@ -1,24 +1,0 @@
-# Documents
-/docs/_*
-/docs/*.log
-
-# Cache
-*__pycache__*
-
-# Debug logs
-/debug/*
-
-# Builds
-/build/*
-/dist/*
-*egg-info*
-
-# Settings
-/setup.cfg
-
-# Star catalogue
-/BSC5*
-
-# Databases except the default
-*.npz
-!/default_database.npz
