--- tetra3/__init__.py
@@ -1,5 +1,0 @@
-name = "tetra3"
-
-from .tetra3 import Tetra3, get_centroids_from_image, crop_and_downsample_image
-
-__all__ = ['Tetra3', 'get_centroids_from_image', 'crop_and_downsample_image']
