--- tetra3/docs/api_docs.rst
@@ -1,21 +1,0 @@
-API reference
-=============
-
-This reference is auto-generated from the source code in the
-`tetra3 GitHub repository <https://github.com/esa/tetra3>`_. General instructions are avilable at
-the `tetra3 ReadTheDocs website <https://tetra3.readthedocs.io/en/latest/>`_.
-
-.. automodule:: tetra3
-
-   tetra3.Tetra3
-   ^^^^^^^^^^^^^
-   .. autoclass:: Tetra3  
-      :members:
-
-   tetra3.get_centroids_from_image
-   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-   .. automethod:: tetra3::get_centroids_from_image
-
-   tetra3.crop_and_downsample_image
-   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-   .. automethod:: tetra3::crop_and_downsample_image
