--- tetra3/docs/index.rst
@@ -1,15 +1,0 @@
-.. tetra3 documentation master file, created by
-   sphinx-quickstart on Wed Jun 19 11:22:42 2019.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-.. include:: ../README.rst
-
-.. toctree::
-   :maxdepth: 3
-   :caption: Contents:
-   
-   self
-   installation
-   api_docs
-
