--- tetra3/requirements.txt
@@ -1,3 +1,0 @@
-numpy==1.21.1
-Pillow==8.3.1
-scipy==1.7.1
